+++ extensions/hypersonic/src/features/onboarding/components/authorization-action.tsx
@@ -1,9 +1,9 @@
-import { useToken } from '@/hooks/use-token'
+import { useAuth } from '@/features/auth/use-auth'
 import { Action, Icon } from '@raycast/api'
 import { authorize } from '@/services/notion/oauth/authorize'
 
 export function AuthorizationAction() {
-  const { actions } = useToken()
+  const { actions } = useAuth()
 
   const handleAuthorize = async () => {
     const token = await authorize()
@@ -18,7 +18,7 @@ export function AuthorizationAction() {
       icon={Icon.Person}
       title={'Authorize'}
       onAction={handleAuthorize}
-      shortcut={{ modifiers: ['cmd'], key: 's' }}
+      shortcut={{ modifiers: ['cmd', 'shift'], key: 'a' }}
     />
   )
 }