+++ extensions/hypersonic/CHANGELOG.md
@@ -1,6 +1,18 @@
 # Hypersonic Changelog
 
-## [0.0.2] - 2022-07-19
+## [0.0.3] - 2022-09-15
+
+- Update Raycast to "@raycast/api": "1.39.2"
+- Update Raycast to "@raycast/utils": "1.4.0"
+- Update Notion to "@notionhq/client": "2.2.0"
+- Major refactor of the code using the new useCachedPromise hook.
+- Added support for custom properties database names.
+- Added support to complete tasks using the new status property from notion.
+- Added support for In progress status when using the new status property.
+- Better integration with notion fetching todos created directly from there without date and name.
+- Change authorize command to `cmd + shift + A`.
+
+## [0.0.2] - 2022-07-15
 
 - Added custom reminder. Now You can set custom reminder for each task.
 - Added filter by label.